feat: better makefile/tox
The make and tox targets that ship with the cookiecutter are a real
mess. Here, we reorganise these targets as follows:

- tox depends on make commands
- make commands do not depend on tox
- make commands are more granular, and faster to run.

In addition, requirement files are reorganised to avoid circular
dependencies and duplicate requirements.

Because mypy checks are now run as part of "tox -e quality", we no
longer need the mypy github action.

Moreover, we introduce the compile-requirements make target, which is
useful for compiling requirements without upgrading them.

We removed the diff_cover make target, which wasn't working anyway
because it defaulted to "--compare-branch=main", and the "main" branch
does not exist.