chore: refactor MongoBackend to simplify mocking
The MongoBackend now provides access only to a database, and it is the
responsibility of each child model to access the collection. This is
done transparently via the _collection property.

The MongoBackend constructor was also simplified. Because we are now
using a single database object, we can drastically simplify mocking in
tests. That's because we don't have to mock models one by one.

And it turns out that we don't need to store the client attribute. Neither do we
need to store a diffent collection name for every model instance, so we
can simply store it as a class attribute. This means that we no longer
need constructors for models.