fix: meilisearch document deletion
In production, we sometimes observe the following error on document
deletion:

    	lms-1  |   File "/openedx/venv/lib/python3.11/site-packages/forum/handlers.py", line 115, in handle_deletion
	lms-1  |     search_backend.delete_document(sender.index_name, document_id)
	lms-1  |   File "/openedx/venv/lib/python3.11/site-packages/forum/search/meilisearch.py", line 93, in delete_document
	lms-1  |     doc_pk = m.id2pk(doc_id)
	lms-1  |              ^^^^^^^^^^^^^^^
	lms-1  |   File "/openedx/venv/lib/python3.11/site-packages/search/meilisearch.py", line 362, in id2pk
	lms-1  |     return hashlib.sha1(value.encode()).hexdigest()
	lms-1  |                         ^^^^^^^^^^^^
	lms-1  | AttributeError: 'int' object has no attribute 'encode'

This is due to the fact that the doc_id passed to the delete_document
method is sometimes an integer, sometimes a str. To reflect that, we
updated the base function signature. We also improved the unit test
coverage. Not all functions are covered yet, but it's getting better.

Note that to reproduce this issue we should use MySQL as a storage
backend.