fix: delete read_states of thread on its deletion
- Management command throws an error for deleted threads
whose read_states still exists in the Users collection. Although it's
a replica of V1(cs_comment_service) which was also keeping the read_states
of deleted threads. This commit will delete the read_states of the thread
from User Collection on thread deletion as keeping the read_states are of no use.
- add unit test for this change