fix: forum bug fixes related to data migration
Bugs:
- No.1: run migration command to migrate data from mongo to mysql,
it doesn't migrate sk(sorting_key) and depth of the sub/child comments.
Due to which sub-comments although being created in mysql doesn't appear
on the frontend. So to reproduce this bug, simple migrate a course which
has a thread with child comments at depth 1
.i.e thread -> comment on the thread -> comment on the parent comment.
Now, migrate this course to mysql. Sub comments(comment on the parent comment)
will not appear on frontend. One can also view the Comment in Comment model
in django admin panel, sk and depth will not be there.

- No.2: create a thread in mongodb, and migrate the course with this thread
to mysql. Now create a comment on this thread in mongo db. Creating a comment
on thread will update it's read_states in User collection and now try to
again migrate it. It will throw an error for last_read_time while updating the
read_states for the user.

This commit fixes above 2 bugs and also incude unit tests.